# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ISC Mbujimayi Digital Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /enrollments:
    get:
      operationId: listEnrollments
      tags: [enrollments]
      summary: List enrollments for current student
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of enrollments with progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollmentListResponse"
    post:
      operationId: createEnrollment
      tags: [enrollments]
      summary: Enroll student in a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnrollmentBody"
      responses:
        "201":
          description: Created enrollment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /enrollments/{id}:
    get:
      operationId: getEnrollmentById
      tags: [enrollments]
      summary: Get enrollment with full progress details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Enrollment with progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollmentDetail"

  # Progress
  /chapters/{chapterId}/progress:
    post:
      operationId: markChapterProgress
      tags: [progress]
      summary: Mark chapter as completed or update video progress
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkProgressBody"
      responses:
        "200":
          description: Progress updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressResult"

  # Payments
...
components:
  schemas:
...
    Filiere:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        description:
          type: ["string", "null"]
        duration:
          type: integer
        studentCount:
          type: integer
        courseCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, code, duration, studentCount, courseCount, createdAt]
...
    Teacher:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        code:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        specialty:
          type: ["string", "null"]
        grade:
          type: ["string", "null"]
        courseCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, userId, code, firstName, lastName, courseCount, createdAt]
...
    Course:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        thumbnail:
          type: ["string", "null"]
        teacherId:
          type: string
        teacher:
          $ref: "#/components/schemas/Teacher"
        filiereId:
          type: ["string", "null"]
        filiere:
          oneOf:
            - $ref: "#/components/schemas/Filiere"
            - type: "null"
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
        level:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        enrollmentCount:
          type: integer
        moduleCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, description, teacherId, teacher, status, enrollmentCount, moduleCount, createdAt, updatedAt]
...
    Enrollment:
      type: object
      properties:
        id:
          type: string
        studentId:
          type: string
        courseId:
          type: string
        course:
          $ref: "#/components/schemas/Course"
        enrolledAt:
          type: string
          format: date-time
        progressPercent:
          type: integer
        completedAt:
          type: ["string", "null"]
          format: date-time
      required: [id, studentId, courseId, course, enrolledAt, progressPercent]
    EnrollmentDetail:
      allOf:
        - $ref: "#/components/schemas/Enrollment"
        - type: object
          properties:
            chapterProgress:
              type: array
              items:
                $ref: "#/components/schemas/ChapterProgress"
          required: [chapterProgress]
    EnrollmentListResponse:
      type: object
      properties:
        enrollments:
          type: array
          items:
            $ref: "#/components/schemas/Enrollment"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
      required: [enrollments, total, page, pageSize, totalPages]
    CreateEnrollmentBody:
      type: object
      properties:
        courseId:
          type: string
      required: [courseId]
    ChapterProgress:
      type: object
      properties:
        id:
          type: string
        chapterId:
          type: string
        completedAt:
          type: ["string", "null"]
          format: date-time
        watchedSeconds:
          type: ["integer", "null"]
      required: [id, chapterId]
    MarkProgressBody:
      type: object
      properties:
        enrollmentId:
          type: string
        completed:
          type: boolean
        watchedSeconds:
          type: integer
      required: [enrollmentId]
    ProgressResult:
      type: object
      properties:
        chapterProgress:
          $ref: "#/components/schemas/ChapterProgress"
        courseProgressPercent:
          type: integer
        certificateGenerated:
          type: boolean
        certificate:
          oneOf:
            - $ref: "#/components/schemas/Certificate"
            - type: "null"
      required: [chapterProgress, courseProgressPercent, certificateGenerated]

    # Payments
...
    Certificate:
      type: object
      properties:
        id:
          type: string
        studentId:
          type: string
        courseId:
          type: string
        course:
          $ref: "#/components/schemas/Course"
        issuedAt:
          type: string
          format: date-time
        hash:
          type: string
        pdfUrl:
          type: ["string", "null"]
      required: [id, studentId, courseId, course, issuedAt, hash]
...